Refactor #182 - Move `Schedule`  to `ui`
