+++ app/src/main/java/com/nlab/reminder/core/schedule/ui/DefaultScheduleElementItemMovePolicy.kt
@@ -1,12 +1,12 @@
 /*
  * Copyright (C) 2024 The N's lab Open Source Project
- *  
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *  
+ *
  *        http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,18 +14,13 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.schedule.view
+package com.nlab.reminder.core.schedule.ui
 
 import com.nlab.reminder.core.schedule.model.ScheduleElement
 
 /**
  * @author thalys
  */
-abstract class ScheduleElementItemMovePolicy {
-    protected abstract fun isMovable(
-        from: ScheduleElement,
-        to: ScheduleElement
-    ): Boolean
-
-    internal fun isMovableInternal(from: ScheduleElement, to: ScheduleElement): Boolean = isMovable(from, to)
+class DefaultScheduleElementItemMovePolicy : ScheduleElementItemMovePolicy() {
+    override fun isMovable(from: ScheduleElement, to: ScheduleElement): Boolean = true
 }
