+++ app/src/main/java/com/nlab/reminder/core/schedule/ui/ScheduleItemViewTypeDelegate.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.schedule.view
+package com.nlab.reminder.core.schedule.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
