--- app/src/main/java/com/nlab/reminder/core/schedule/view/DragPositionHolder.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright (C) 2024 The N's lab Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nlab.reminder.core.schedule.view
-
-/**
- * @author thalys
- */
-internal class DragPositionHolder {
-    private var from: Int? = null
-    private var to: Int? = null
-
-    /**
-     * @return first is from, second is to
-     */
-    fun snapshot(): Pair<Int, Int>? {
-        val curFrom: Int = from ?: return null
-        val curTo: Int = to ?: return null
-        return curFrom to curTo
-    }
-
-    fun setPosition(from: Int, to: Int) {
-        if (this.from == null) this.from = from
-        this.to = to
-    }
-
-    fun clearPosition() {
-        from = null
-        to = null
-    }
-}
